feat: support a flag to enable the legacy block prop (#2518)
Backports https://github.com/celestiaorg/celestia-core/pull/2512 to main

+++ consensus/propagation/reactor_test.go
@@ -51,7 +51,7 @@ func newPropagationReactor(s *p2p.Switch, tr trace.Tracer, pv types.PrivValidato
 		Mempool:       &mockMempool{txs: make(map[types.TxKey]*types.CachedTx)},
 		Privval:       pv,
 		ChainID:       TestChainID,
-		BlockMaxBytes: types.MaxBlockSizeBytes,
+		BlockMaxBytes: int64(types.MaxBlockSizeBytes),
 	})
 	blockPropR.traceClient = tr
 	blockPropR.currentProposer = pub
